🎉 Change strategy for updating WebView
Previously, the strategy involved reflecting HTML changes on the WebView back to the source code and then reflecting those changes again on the WebView.
To keep element selection on the WebView, it was necessary to embed the selection state in the HTML.
To avoid updating the source code just by selecting elements, the source code range update upon HTML changes on the WebView is now performed within the WebView itself.